requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
gunicorn==23.0.0
Jinja2==3.1.5
Werkzeug==3.1.3
itsdangerous==2.2.0
click==8.1.8
python-dotenv==1.0.1
affinidi_tdk_auth_provider==1.33.2
affinidi_tdk_credential_issuance_client==1.64.0
pypdf==5.1.0
affinidi_tdk_credential_verification_client==1.57.0
requests==2.32.3
reportlab==4.2.5
cryptography==44.0.0
qrcode==8.0
affinidi_tdk_wallets_client==1.58.0
PyJWT==2.10.1
affinidi_tdk_iota_client==1.56.0
Authlib==1.6.1